Update README.md
Thia exploratory analysis shows how borrowers relate with the loan process. This analysis will determine, a major decision by the firm of categories of borrowers to prioritize loan, and how the returns affect investor' motivation also.
Key Insights for Presentation
The focus is on features that  affect the borrower APR, which are original loan amount and duration, Prosper rating. I started by showing the distribution of number of borrowers at various percentage rates, to depict what loan term is most apealing to borrowers, this is to eable the organisation to priotize the category to consolidate in order to have a better turn over, as every Organisation continue to stay in business when there is a better structure, and this analysis will better direct the orginsation on best decisions to make.